lint

+++ mount/utils/transform/getattr.go
@@ -82,7 +82,7 @@ func unixFSAttr(ufsNode *unixfs.FSNode, req IPFSStatRequest) (*IPFSStat, IPFSSta
 		// NOTE: we can't account for variable block size so we use the size of the first block only (if any)
 		blocks := len(ufsNode.BlockSizes())
 		if blocks > 0 {
-			attr.BlockSize = uint64(ufsNode.BlockSize(0))
+			attr.BlockSize = ufsNode.BlockSize(0)
 			attr.Blocks = uint64(blocks)
 		}
 		// 0 is a valid value for these fields, especially for non-regular files
